chore: move to i18n utils file
